examples/sdks/python/pyproject.toml
[project]
name = "latitude-python-examples"
version = "1.0.0"
requires-python = ">=3.9, <3.13"
dependencies = [
    "openai>=1.58.1",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.1",
    "latitude-sdk",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.hatch.metadata]
allow-direct-references = true

[tool.uv.sources]
latitude-sdk = { path = "../../../packages/sdks/python" }

